requirements.txt
streamlit==1.39.0
requests>=2.31.0
pandas>=2.2.0
sendgrid>=6.11.0
apscheduler>=3.10.0
supabase>=2.0.0
python-dotenv>=1.0.0